requirements.txt
python-multipart
fastapi
uvicorn
pytesseract
opencv-python
Pillow
transformers
torch
stanza
mysql-connector-python
numpy
pandas
pydantic
pydantic-core
pillow
opencv-python
pytesseract
python-dateutil
requests
scikit-learn
python-docx
numpy
scipy
pandas
openpyxl